+++ Meta/gn/secondary/Userland/Libraries/LibWeb/CSS/StyleValues/BUILD.gn
@@ -11,9 +11,9 @@ source_set("StyleValues") {
     "CSSHSL.cpp",
     "CSSHWB.cpp",
     "CSSKeywordValue.cpp",
+    "CSSLabLike.cpp",
     "CSSMathValue.cpp",
     "CSSOKLCH.cpp",
-    "CSSOKLab.cpp",
     "CSSRGB.cpp",
     "ConicGradientStyleValue.cpp",
     "ContentStyleValue.cpp",
